# tests/golden/contracts/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Greenhouse API",
  "description": "Smart plant irrigation system with evidence-based plant care, multi-sensor conflict resolution, and self-learning irrigation profiles.\n\n",
  "version": "1.0.0"
 },
 "paths": {
...
  "/api/v1/clusters/{cluster_id}/plants": {
   "post": {
    "tags": [
     "plants"
    ],
    "summary": "Add Plant",
    "description": "Add a plant to a cluster.\n\nCare thresholds (water needs, temperature/humidity ranges) can be supplied\ndirectly or left null to be filled in later via POST /plants/sync, which\nlooks them up in the evidence-based plant database.\n\nArgs:\n    cluster_id: ID of the cluster the plant belongs to.\n    request: Species and optional plant-care fields.\n\nRaises:\n    HTTPException: 404 if the cluster does not exist.\n\nReturns:\n    The created plant including its ID and parent cluster.",
    "operationId": "add_plant",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "cluster_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Cluster Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/CreatePlantRequest"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/PlantResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "get": {
    "tags": [
     "plants"
    ],
    "summary": "List Plants",
    "description": "List every plant in a cluster.\n\nArgs:\n    cluster_id: ID of the cluster to enumerate.\n\nReturns:\n    The cluster's plants (empty for an unknown cluster).",
    "operationId": "list_plants",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "cluster_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Cluster Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/PlantResponse"
         },
         "title": "Response List Plants"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/clusters/{cluster_id}/plants/{plant_id}": {
   "put": {
    "tags": [
     "plants"
    ],
    "summary": "Update a plant",
    "description": "Partially update a plant's care metadata.\n\nOnly fields present in the request body are modified; omitted fields are\nleft unchanged. The plant must belong to the specified cluster.\n\nArgs:\n    cluster_id: Cluster the plant belongs to.\n    plant_id: Numeric plant identifier.\n    request: Fields to update — any subset of care metadata fields.\n\nReturns:\n    The updated plant.\n\nRaises:\n    HTTPException: 404 if the plant does not exist or belongs to a\n        different cluster.",
    "operationId": "update_plant",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "cluster_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Cluster Id"
      }
     },
     {
      "name": "plant_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Plant Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/UpdatePlantRequest"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/PlantResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "delete": {
    "tags": [
     "plants"
    ],
    "summary": "Delete a plant",
    "description": "Delete a plant from a cluster.\n\nSensors previously linked to this plant retain their cluster membership\nbut have their plant_id set to null. This operation is irreversible.\n\nArgs:\n    cluster_id: Cluster the plant belongs to.\n    plant_id: Numeric plant identifier.\n\nReturns:\n    success=True on successful deletion.\n\nRaises:\n    HTTPException: 404 if the plant does not exist or belongs to a\n        different cluster.",
    "operationId": "delete_plant",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "cluster_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Cluster Id"
      }
     },
     {
      "name": "plant_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Plant Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SuccessResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/vacation": {
   "get": {
    "tags": [
     "vacation"
    ],
    "summary": "List vacation windows",
    "description": "Return all vacation windows together with the currently active one.\n\nReturns:\n    A list of all windows and the active window (spanning now), if any.",
    "operationId": "list_vacation_windows",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/VacationListResponse"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   },
   "post": {
    "tags": [
     "vacation"
    ],
    "summary": "Create a vacation window",
    "description": "Schedule a vacation window that makes the engine ration water to last the trip.\n\nWhile the window is active the decision engine appends a\n``VACATION_ACTIVE`` reason to every decision and, for clusters whose\nirrigators have both ``reservoir_l`` and ``flow_rate_l_per_min`` set,\nenforces a per-day burn-down budget: each irrigation's duration is trimmed\nto fit the remaining tank allowance (``VACATION_RATIONING``), or flipped to\nSKIP when no usable water remains this cycle (``VACATION_BUDGET_EXHAUSTED``).\nWhen no irrigator in a cluster has capacity configured the window is purely\ninformational and irrigation proceeds exactly as it would normally.\n\nArgs:\n    request: Window start and end as Unix timestamps, optional contact email\n        and free-text notes.\n\nReturns:\n    The newly created vacation window.\n\nRaises:\n    HTTPException: 400 if ``starts_at`` is not strictly before ``ends_at``.",
    "operationId": "create_vacation_window",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/VacationCreateRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/VacationResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/v1/vacation/{window_id}": {
   "put": {
    "tags": [
     "vacation"
    ],
    "summary": "Update a vacation window",
    "description": "Partially update a vacation window.\n\nOnly fields present in the request body are modified; omitted fields are\nleft unchanged. The effective ``starts_at`` and ``ends_at`` after the patch\nmust satisfy ``starts_at < ends_at`` — otherwise the request is rejected\nso the vacation gate can never end up reversed.\n\nArgs:\n    window_id: Numeric ID of the vacation window to update.\n    request: Any subset of starts_at, ends_at, contact_email, notes.\n\nReturns:\n    The updated vacation window.\n\nRaises:\n    HTTPException: 404 if no window with that ID exists, 400 if the\n        resulting ``starts_at`` is not strictly before ``ends_at``.",
    "operationId": "update_vacation_window",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "window_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Window Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/UpdateVacationWindowRequest"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/VacationResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "delete": {
    "tags": [
     "vacation"
    ],
    "summary": "Delete a vacation window",
    "description": "Remove a vacation window by ID.\n\nArgs:\n    window_id: Numeric ID of the vacation window to delete.\n\nReturns:\n    Success acknowledgement.\n\nRaises:\n    HTTPException: 404 if no window with that ID exists.",
    "operationId": "delete_vacation_window",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "window_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Window Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SuccessResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "CreatePlantRequest": {
    "properties": {
     "species": {
      "type": "string",
      "title": "Species"
     },
     "category": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Category"
     },
     "water_needs": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Water Needs"
     },
     "light_needs": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Light Needs"
     },
     "ideal_temp_min": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Ideal Temp Min"
     },
     "ideal_temp_max": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Ideal Temp Max"
     },
     "ideal_humidity_min": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Ideal Humidity Min"
     },
     "ideal_humidity_max": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Ideal Humidity Max"
     },
     "notes": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Notes"
     }
    },
    "type": "object",
    "required": [
     "species"
    ],
    "title": "CreatePlantRequest",
    "description": "Body for adding a plant to a cluster."
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "PlantResponse": {
    "properties": {
     "species": {
      "type": "string",
      "title": "Species"
     },
     "category": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Category"
     },
     "water_needs": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Water Needs"
     },
     "light_needs": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Light Needs"
     },
     "ideal_temp_min": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Ideal Temp Min"
     },
     "ideal_temp_max": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Ideal Temp Max"
     },
     "ideal_humidity_min": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Ideal Humidity Min"
     },
     "ideal_humidity_max": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Ideal Humidity Max"
     },
     "notes": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Notes"
     },
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "cluster_id": {
      "type": "integer",
      "title": "Cluster Id"
     }
    },
    "type": "object",
    "required": [
     "species",
     "id",
     "cluster_id"
    ],
    "title": "PlantResponse",
    "description": "A plant as stored, with the cluster it belongs to."
   },
...
   "SuccessResponse": {
    "properties": {
     "success": {
      "type": "boolean",
      "title": "Success"
     }
    },
    "type": "object",
    "required": [
     "success"
    ],
    "title": "SuccessResponse",
    "description": "Generic success acknowledgement."
   },
...
   "UpdatePlantRequest": {
    "properties": {
     "species": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Species"
     },
     "category": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Category"
     },
     "water_needs": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Water Needs"
     },
     "light_needs": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Light Needs"
     },
     "ideal_temp_min": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Ideal Temp Min"
     },
     "ideal_temp_max": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Ideal Temp Max"
     },
     "ideal_humidity_min": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Ideal Humidity Min"
     },
     "ideal_humidity_max": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Ideal Humidity Max"
     },
     "notes": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Notes"
     }
    },
    "type": "object",
    "title": "UpdatePlantRequest",
    "description": "Partial update of a plant; omitted fields keep their value."
   },
...
   "UpdateVacationWindowRequest": {
    "properties": {
     "starts_at": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Starts At"
     },
     "ends_at": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Ends At"
     },
     "contact_email": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Contact Email"
     },
     "notes": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Notes"
     }
    },
    "type": "object",
    "title": "UpdateVacationWindowRequest",
    "description": "Partial-update body for a vacation window. All fields optional."
   },
   "VacationCreateRequest": {
    "properties": {
     "starts_at": {
      "type": "integer",
      "title": "Starts At"
     },
     "ends_at": {
      "type": "integer",
      "title": "Ends At"
     },
     "contact_email": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Contact Email"
     },
     "notes": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Notes"
     }
    },
    "type": "object",
    "required": [
     "starts_at",
     "ends_at"
    ],
    "title": "VacationCreateRequest",
    "description": "Body for scheduling a vacation window (Unix seconds); watering is rationed inside it."
   },
   "VacationListResponse": {
    "properties": {
     "active": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/VacationResponse"
       },
       {
        "type": "null"
       }
      ]
     },
     "items": {
      "items": {
       "$ref": "#/components/schemas/VacationResponse"
      },
      "type": "array",
      "title": "Items"
     }
    },
    "type": "object",
    "required": [
     "active",
     "items"
    ],
    "title": "VacationListResponse",
    "description": "All vacation windows and the one active now, if any."
   },
   "VacationResponse": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "starts_at": {
      "type": "integer",
      "title": "Starts At"
     },
     "ends_at": {
      "type": "integer",
      "title": "Ends At"
     },
     "contact_email": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Contact Email"
     },
     "notes": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Notes"
     },
     "created_at": {
      "type": "integer",
      "title": "Created At"
     }
    },
    "type": "object",
    "required": [
     "id",
     "starts_at",
     "ends_at",
     "contact_email",
     "notes",
     "created_at"
    ],
    "title": "VacationResponse",
    "description": "A vacation window as stored."
   },
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  },
  "securitySchemes": {
   "HTTPBearer": {
    "type": "http",
    "scheme": "bearer"
   }
  }
 },
...
}